Create challenge6.rb
challenge 6:
Today's challenges:

Beginner:
There are four variables to handle in this challenge. A start and an end point, as well as two other integers a and b (you can have them as set variables that you hard code in, user input or parameters to a function). Write code that prints (or returns if it’s a function) all the numbers between the start and the end point (both included) that are multiples of a but not divisible by b. The start point, the end point, a and b are all integers.
Example: if start = 10, end = 20, a = 2 and b = 4 → print 10, 14, and 18, because they are the only numbers between 10 and 20 that are multiples of 2 but not divisible by 4.

Advanced:
Write a function that takes two arguments a (string) and b (list of lists). String a will have a dash (-) instead of certain letters, and list b will contain lists of length 2, containing a string of a letter and an integer. The integer represents the order in which you replace the letters.
Example: a = “h-l--” and b = [[“o”, 3], [“e”, 1], [“l”, 2]] → “hello”
Note that the integer accompanying the string in the lists represents the order in which you insert the letters, not the index at which you insert them.

Good luck to all of you!