Challenge 7: Beginner: Write code that takes in a string and returns/prints the string but with a capitalized letter at the beginning of each word. Use logic, not a built in function. Note that the string given is always in full lowercase. Example: given_string = “hello world i am doing great!”, print/return “Hello World I Am Doing Great!”
Advanced:
Create a function that takes two arguments: an array and a number. In the array (the first argument), if an element occurs more than n times (the second argument), remove the extra occurrence(s) and return the result.
Example: list_given = [1, 1, 2, 3, 1, 3], n = 2 → [1, 1, 2, 3, 3], because 1 appears more than 2 times, so its extra occurrence is removed.